requirements.txt
streamlit==1.39.0
langchain==0.0.284
openai==0.28.1
faiss-cpu==1.8.0.post1
python-dotenv==1.0.1